twitter_clone_api/requirements.txt
Django>=3.0,<4.0
grpcio>=1.36.1
grpcio-tools>=1.36.1
djangorestframework>=3.12.4
pytest>=6.1.1